Set up exception handling logout endpoint

+++ server/src/main/java/server/handlers/LogoutHandler.java
@@ -1,5 +1,9 @@
 package server.handlers;
 
+import java.util.Map;
+
+import com.google.gson.Gson;
+
 import service.UserService;
 import spark.Request;
 import spark.Response;
@@ -12,9 +16,23 @@ public class LogoutHandler {
     }
 
     public Object handle(Request req, Response res) {
-        final var authToken = req.headers("Authorization");
-        userService.deleteAuth(authToken);
-        res.status(200);
-        return "";
+        try {
+            final var authToken = req.headers("Authorization");
+            if (authToken == null) {
+                res.status(401);
+                return new Gson().toJson(Map.of("message", "Error: unauthorized"));
+            }
+            final var auth = userService.getAuth(authToken);
+            if (auth == null) {
+                res.status(401);
+                return new Gson().toJson(Map.of("message", "Error: unauthorized"));
+            }
+            userService.deleteAuth(authToken);
+            res.status(200);
+            return "";
+        } catch(Exception e) {
+            res.status(500);
+            return "Error: Internal server error";
+        }
     }
 }
